So... hear me out:
what if you wanted to remove A BUNCH OF WOOD in a dark oak forest, because you like the Grass Color, but not the Trees? Or in a Spruce Forest? Well look no further than the newest invention of the crazy mind that is @schnow265 : The **TreeStroyer**! It will remove the selected piece of wood in a Radius of 350 Blocks! All of that for the price of:

- 1 x Choppy (6 Amethyst Shards and one Stick
- 4 x Netherite Blocks
- 2 x Nether Star
- 1 x Ruby
- 1 x Elytra